requirements.txt
ensure==1.0.2
fastapi
pydantic
requests
uvicorn
mkdocs-material
python-dotenv
pandas
-e .s